backend/ingestion/infra_poller/pyproject.toml
[project]
name = "sovereign-watch-infra-poller"
version = "0.1.0"
description = "Sovereign Watch infrastructure ingestion poller"
requires-python = ">=3.12"
dependencies = [
    "requests==2.32.5",
    "redis==7.3.0",
    "psycopg2-binary>=2.9.9",
]
